UPSTREAM: timekeeping: Add missing _ns functions for coarse accessors
This further unifies the accessors for the fast and coarse functions, so
that the same types of functions are available for each. There was also
a bit of confusion with the documentation, which prior advertised a
function that has never existed. Finally, the vanilla ktime_get_coarse()
was omitted from the API originally, so this fills this oversight.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Arnd Bergmann <arnd@arndb.de>
Link: https://lkml.kernel.org/r/20190621203249.3909-3-Jason@zx2c4.com
(cherry picked from commit 4c54294d01e605a9f992361b924c5d8b12822a6d)

BUG=b:169699158
TEST=Build

Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I1a467d862f282defb537d5d8a17d3bfbe750a8c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2547624
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>