CHROMIUM: iwl7000: mvm: fix the reordering buffer loose sync
I looked at the NSSN where I should have looked at the SN.
The NSSN can go forward in big steps and hence we are not
guaranteed that it'll be exactly 0 / 2048 at any point in
time.

Also we should sync the head_sn of all the queues when
we have an NSSN lagging behind because of a packet that
timed out. If that happened, and we now have an empty
reordering buffer, we can safely release frames based
on our head_sn (this is the whole point of the head_sn)
and not wait for the NSSN to catch up.  In this flow, we
will release frames, and those frames may be 0 or 2048.

Change-Id: I21edc58a1becc4371ae6c0fdd39bf8470959006d
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: a5fb29a51b7835576724f8bb379ffb328e7b4de2