CHROMIUM: iwl7000: mvm: send LQ command always ASYNC
The only place where the command was sent as SYNC is during
init and this is not really critical. This change is required
for replacing RS mutex with a spinlock (in the subsequent patch),
since SYNC comamnd requres sleeping and thus the flow cannot
be done when holding a spinlock.

[NOTE for Core45]: this patch should have gone in before the other
patches RS patches in the last series.  Not having it caused something
to break, so apply it now.

Change-Id: I33f8e4fa26d098794a78eb54d874b0eee8680e26
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: a442a79d384781d1dcdc5a8cfd243e92499dacd9