CHROMIUM: iwl7000: mvm: send LQ command always ASYNC
The only place where the command was sent as SYNC is during
init and this is not really critical. This change is required
for replacing RS mutex with a spinlock (in the subsequent patch),
since SYNC comamnd requres sleeping and thus the flow cannot
be done when holding a spinlock.

Change-Id: I33f8e4fa26d098794a78eb54d874b0eee8680e26
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
iwl7000-tree: ac8e286ecebed2206a27602080f322860c249aa3