BACKPORT: iommu: Introduce struct iommu_iotlb_gather for batching TLB flushes
To permit batching of TLB flushes across multiple calls to the IOMMU
driver's ->unmap() implementation, introduce a new structure for
tracking the address range to be flushed and the granularity at which
the flushing is required.

This is hooked into the IOMMU API and its caller are updated to make use
of the new structure. Subsequent patches will plumb this into the IOMMU
drivers as well, but for now the gathering information is ignored.

Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit a7d20dc19d9ea7012227be5144353012ffa3ddc4)
[Context conflicts in include/linux/iommu.h]

BUG=b:139716453
TEST=Boot kukui

Change-Id: I3145a28c3660bca9a9fc90c60005b272697fefeb
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1833380
Reviewed-by: Tomasz Figa <tfiga@chromium.org>