CHROMIUM: iwl7000: mvm: fix endiannes in ppag API and other clean-ups
We were using u32 in the iwl_ppag_table_cmd structure, but it's a FW
API, so it must use little-endian.  Fix that.

And while at it, remove the mvm version of the struct, since it's
pretty much the same as the one used by the FW API.  We can save the
actual command to send already when reading and storing.

Additionally, fix the table revision check.  We should only support
revision 0.  And fix the check for !enabled, since we were checking if
the ACPI element was !NULL instead of checking the actual value.

Change-Id: Idbbcffa272a41e849050206edb538131e8a7f3f5
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: ef1fa28b9db96b15ab91f9ca77c55515e4842aa2