Revert "FROMLIST: mmc: dw_mmc: Disable SDIO interrupts while suspended to fix suspend/resume"
This reverts commit 2769711eafcea5e7828803a188d5ca3d2f8accb1.

While the patch works OK and doesn't hurt, upstream solved it in a
different way and we already have that other way since it came in
through stable.  See commit 98467b8fda41 ("mmc: core: Prevent
processing SDIO IRQs when the card is suspended").

Let's do this revert so it's clear that it's not needed for the next
rebase.

BUG=chromium:941638
TEST=suspend_stress_test

Change-Id: Ie9d8746485b019372005dbfe0e3fbe258826a2cb
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1709940
Reviewed-by: Guenter Roeck <groeck@chromium.org>