CHROMIUM: iwl7000: mvm: fix a use-after-free bug in iwl_mvm_tx_tso_segment
Accessing the hdr of an skb that was consumed already isn't
a good idea.
First ask if the skb is a QoS packet, then keep that data
on stack, and then consume the skb.
This was spotted by KASAN.

Change-Id: Ia0c41e9cad6da74019f4f5b3114559bf672c3a9b
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: b178ef44801cdb77a230e5886ee25ad3fef18a87