BACKPORT: FROMLIST: clk: Remove cached cores in parent map during unregister
As clocks are registered their parents are resolved and the parent_map
is updated to cache the clk_core objects of each existing parent.
But in the event of a clock being unregistered this cache will carry
dangling pointers if not invalidated, so do this for all children of the
clock being unregistered.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(am from https://lore.kernel.org/linux-arm-msm/20190723051446.20013-1-bjorn.andersson@linaro.org/)

Downstream changes:
- downstream parents array is already clk_core

BUG=None
TEST=Boot cheza and no kernel oops

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Id0c742eede0c739a46506aa431130d0e685d147a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1744637
Reviewed-by: Stephen Boyd <swboyd@chromium.org>