FROMLIST: sched/fair: core wide vruntime comparison
This patch provides a vruntime based way to compare two cfs task's
priority, be it on the same cpu or different threads of the same core.

When the two tasks are on the same CPU, we just need to find a common
cfs_rq both sched_entities are on and then do the comparison.

When the two tasks are on differen threads of the same core, the root
level sched_entities to which the two tasks belong will be used to do
the comparison.

An ugly illustration for the cross CPU case:

   cpu0         cpu1
 /   |  \     /   |  \
se1 se2 se3  se4 se5 se6
    /  \            /   \
  se21 se22       se61  se62

Assume CPU0 and CPU1 are smt siblings and task A's se is se21 while
task B's se is se61. To compare priority of task A and B, we compare
priority of se2 and se6. Whose vruntime is smaller, who wins.

To make this work, the root level se should have a common cfs_rq min
vuntime, which I call it the core cfs_rq min vruntime.

When we adjust the min_vruntime of rq->core, we need to propgate
that down the tree so as to not cause starvation of existing tasks
based on previous vruntime.

BUG=b:152605392
TEST=run power_VideoCall test

(am from https://lore.kernel.org/lkml/2f83d888890cec14be3a7aead0859dceebb4012f.1583332765.git.vpillai@digitalocean.com/)

Signed-off-by: Aaron Lu <ziqian.lzq@antfin.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: Ibf44492085ff31f376b529b0b94034f6bdb31895
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2197805