CHROMIUM: media: uapi: Restore v4l2_ctrl_h264_slice_params to what Chrome expects
Since the UAPI is still in staging and not stable, we ended up with a
change that broke the compatibility, but also added some other
definitions that we don't want to remove. Rather than reverting it, just
manually bring the struct back to its original state.

BUG=chromium:965375
BUG=chromium:992295
TEST=emerge-veyron_minnie-kernelnext chromeos-kernel-4_19

Change-Id: I1e17e1f23f2b0e6c688f0f26c5138d1b5df43057
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1746011
Reviewed-by: Francois Buergisser <fbuergisser@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>