CHROMIUM: net: phy: realtek: Add LED configuration support for RTL8211E
Add a .config_led hook which is called by the PHY core when
configuration data for a PHY LED is available. Each LED can be
configured to be solid 'off, solid 'on' for certain (or all)
link speeds or to blink on RX/TX activity.

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
(upstream attempt: https://lore.kernel.org/patchwork/patch/1114555/)

BUG=chromium:941638
TEST=build for veyron

Change-Id: I3716f60a16a74f35373e5410b6b4717da241e94c
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1776789
Reviewed-by: Douglas Anderson <dianders@chromium.org>