CHROMIUM: drm/vgem: plug in our own dma_ops
On arch/arm64, we end up w/ dummy_dma_ops, which has a ->map_sg() that
is hard coded to return 0.  This cases dma-buf imports into vgem to fail
(because dma_map_sg_attrs() fails).

This is no longer needed after v5.0 or so, as the DMA API implementation
on arch/arm64 has been substantially changed with the introduction of
the fast-path for dma-direct.

BUG=chromium:983368
TEST=vgem_test on cheza

Change-Id: I5fc889dd2a5894aba267f02714637517b8d99f82
Signed-off-by: Rob Clark <robdclark@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1754965
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>