CHROMIUM: iwl7000: mac80211: initialize chandef in ieee80211_determine_chantype()
Since the extra data for EDMG was introduced in the v5.4's version of
the cfg80211_chan_def structure, the EDMG part will remain
uninitialized in ieee_determine_chantype() when using older Core52 or
older.  So the chandef may be rejected later on, because the EDMG
channel validity check may fail, depending on the values that remained
in the unitialized part.

This patch brings in the mac80211 part of 2a38075cd0be ("nl80211: Add
support for EDMG channels") and is only needed in Core50 and Core52.

Change-Id: I3b433f8b9338d800bdbb327ed2ec4c60900119c0
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: c5058592ae51a9641f34e56a461811b30e12f51d