FROMGIT: iommu/io-pgtable-arm-v7s: Extend to support PA[33:32] for MediaTek
MediaTek extend the arm v7s descriptor to support up to 34 bits PA where
the bit32 and bit33 are encoded in the bit9 and bit4 of the PTE
respectively. Meanwhile the iova still is 32bits.

Regarding whether the pagetable address could be over 4GB, the mt8183
support it while the previous mt8173 don't, thus keep it as is.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Acked-by: Will Deacon <will@kernel.org>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 4c019de653237674d38cf2b3119153b144ffe173
 git://git.kernel.org/pub/scm/linux/kernel/git/joro/iommu.git arm/mediatek)

BUG=b:126008328
TEST=Boot kukui

Change-Id: Ic954f3550d84eeb11aac8cee69861a18e0dec71a
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1780686
Reviewed-by: Sean Paul <seanpaul@chromium.org>