CHROMIUM: dt-bindings: net: phy: Add subnode for LED configuration
The LED behavior of some Ethernet PHYs is configurable. Add an
optional 'leds' subnode with a child node for each LED to be
configured. The binding aims to be compatible with the common
LED binding (see devicetree/bindings/leds/common.txt).

A LED can be configured to be:

- 'on' when a link is active, some PHYs allow configuration for
  certain link speeds
- 'off'
- blink on RX/TX activity, some PHYs allow configuration for
  certain link speeds

For the configuration to be effective it needs to be supported by
the hardware and the corresponding PHY driver.

Suggested-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
(upstream attempt: https://lore.kernel.org/patchwork/patch/1114553/)

BUG=chromium:941638
TEST=build for veyron

Change-Id: I5d494857fa98eb7e827657a8cb86f4c3fbc6cc5d
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1776785
Reviewed-by: Douglas Anderson <dianders@chromium.org>