UPSTREAM: ASoC: SOF/Intel: fix selection of SND_INTEL_NHTL
We should only select SND_INTEL_NHLT when ACPI is defined. This was
done for the legacy HDAudio driver but not for DSP-enabled cases,
leading to compilation errors with randconfig.

Fix by aligning on the same solution.

For the Skylake driver this is overkill since there is a top-level
dependency on ACPI, but it doesn't hurt and it's better to have
consistency.

Fixes: 68b953aeb50d9 ('ASoC: SOF: Intel: hda: fixup HDaudio topology name with DMIC number')
Reported-by: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20190829214213.11653-1-pierre-louis.bossart@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 37638af8135c44e9c9e4b2744eadfd3839efc023)
Signed-off-by: Ben Zhang <benzh@chromium.org>

BUG=b:140723130
TEST=Test Audio use cases for CML with full SOF patch series applied

Change-Id: Id2d9ef39dab153602668f88fac0f207dbbc59c76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2097609
Tested-by: Ben Zhang <benzh@chromium.org>
Commit-Queue: Ben Zhang <benzh@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Ben Zhang <benzh@chromium.org>