CHROMIUM: media: mtk-vcodec: venc: handle firmware version field
Firmwares for encoders newer than MT8173 will include an ABI version
number in their initialization ack message. Add the capacity to manage
it and make initialization fail if the firmware ABI is of a version that
we don't support.

For MT8173, this ABI version field is reserved and thus undefined ; thus
ignore it on this chip. There should only be one firmware version available
for it anyway.

BUG=b:142687721
TEST=Kernel builds and boots on Kukui.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: I221236d79ea6e00bed292fbff22e0e2951069201
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1781257
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>