CHROMIUM: Bluetooth: btmrvl: Detect hangs and force a replug of the SDIO card
When scanning for BLE devices for a longer period (e.g. because a
BLE device is paired, but not connected) the Marvell 8997 often
ends up in a borked state, which manifests through failures on
certain SDIO transactions.

When such a SDIO failure is detected force a replug of the SDIO
card to initialize it from scratch. Since the SDIO bus is shared
with the WiFi part of the chip this also involves a reset of WiFi.

We probably want to upstream this patch once "mmc: core: Add
sdio_trigger_replug() API" (https://lore.kernel.org/patchwork/patch/1103196/)
or similar lands in the maintainer tree.

BUG=chromium:1000492, chromium:990991
TEST=on jerry with a BLE device paired, but not connected:
  suspend/resume 100x
    => Bluetooth still works
  leave the device running for 12h
    => Bluetooth still works

Change-Id: I40cc9b3d5de04f0631c931d94757fb0f462b24bd
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1803515
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>