CHROMIUM: drm/bridge: dw-hdmi: Add HDCP support
This patch adds support to dw-hdmi to enable HDCP (High-bandwidth
Digital Content Protection).  A few notes about the implementation:

1. The method for enabling HDCP and for communicating HDCP status back
   to userspace is the same as the Intel HDCP driver.  That is:
   userspace indicates Desired/Undesired via the "Content Protection"
   property and the kernel is in charge of changing the same property
   between Desired/Enabled.  This is the method that Chrome on Chrome
   OS uses and (since I see that the Intel driver is upstream) is
   apparently the suggested way to do it upstream.
2. HDCP negotiation isn't fast.  In order to keep things performant,
   transitions from Desired to Enabled are asynchronous.
3. Because both userspace and the kernel are communicating via a
   single shared property, keeping things race free is tricky.  We
   accomplish that in much the same way that the Intel driver does by
   forking much of the job off to a worker.
4. This implementation basically just has our worker polling the DRM
   Property every few seconds.  This is both for simplicity (don't
   need to spend so much code trying to start/stop the worker) and
   because I really am not sure where to hook into DRM.  Maybe it's
   OK, though?

While this patch adds support to dw-hdmi for HDCP, it is missing a
critical piece: the piece to actually get the HDCP key and pass it in.
Specifically a future patch needs to be written that will actually
call dw_hdmi_config_hdcp_key().  Because of this it seems unlikely
that upstream will be willing to land this patch.

This patch was tested with a patch which adds a sysfs interface to
dw_hdmi-rockchip allowing userspace to provide this key titled
("CHROMIUM: drm/rockchip: dw_hdmi: Support Chromebook-style HDCP key
from userspace").  See that patch for some details about why it may
not be appropriate for upstream.

It is possible that there may be SoCs that have the HDCP key burned
into a ROM that dw-hdmi has access to.  If such SoCs exist it is
possible that they could be enabled in a simpler way.

NOTE ALSO: This forward ports / adapts HDCP support from the Chrome OS
3.14 tree.  In the Chrome OS tree all that code is attributed to Yakir
Yang <ykk@rock-chips.com>.  I specifically haven't spent much time
comparing to the code in Rockchip's github tree.

For patches from Chrome OS 3.14:
- https://crrev.com/c/237585
  CHROMIUM: drm: bridge/dw_hdmi: add HDCP support
- https://crrev.com/c/302688
  CHROMIUM: drm: bridge/dw_hdmi: Remove re-enable encryption optimization
- https://crrev.com/c/302689
  CHROMIUM: drm: bridge/dw_hdmi: Only enable "HDCP enabled" interrupts

BUG=chromium:971474
TEST=Use this to generate HDCP seed

Change-Id: I1f4ac511451d7da78a869885e89e0a8bbc0a8b84
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1814424
Reviewed-by: Sean Paul <seanpaul@chromium.org>